+++ src/Compiler/IL/LIRInstructionInfo.cs
@@ -181,6 +181,9 @@ internal static partial class LIRInstructionInfo
         typeof(LIRInOperator),
         typeof(LIRInstanceOfOperator),
         typeof(LIRIsInstanceOf),
+        typeof(LIRPrivateBrandCheck),
+        typeof(LIRLoadPrivateReceiverField),
+        typeof(LIRStorePrivateReceiverField),
         typeof(LIRLabel),
         typeof(LIRLeave),
         typeof(LIRLeftShift),
@@ -430,6 +433,8 @@ internal static partial class LIRInstructionInfo
                 or LIRCopyTemp
                 => LIRInstructionEffects.None,
 
+            LIRPrivateBrandCheck => LIRInstructionEffects.MayThrow,
+
             LIRLoadParameter
                 or LIRLoadThis
                 or LIRLoadScopesArgument
@@ -508,6 +513,7 @@ internal static partial class LIRInstructionInfo
                 or LIRGetStringLength
                 or LIRGetInferredMember
                 or LIRLoadUserClassInstanceField
+                or LIRLoadPrivateReceiverField
                 or LIRLoadUserClassStaticField
                 or LIRGetIntrinsicGlobal
                 or LIRGetIntrinsicGlobalFunction
@@ -522,6 +528,7 @@ internal static partial class LIRInstructionInfo
                 or LIRArrayPushRange
                 or LIRArrayAdd
                 or LIRStoreUserClassInstanceField
+                or LIRStorePrivateReceiverField
                 or LIRStoreUserClassStaticField
                 => HeapWriteEffects,