+++ tests/Jroc.Tests/Classes/JavaScript/Classes_ClassFieldArrowInitializer.js
@@ -1,0 +1,6 @@
+class Example {
+    value = 4;
+    read = () => this.value;
+}
+
+console.log(new Example().read());